Separate singleton metaclass to use globally.
