+++ fukurou/patterns/__init__.py
@@ -1,0 +1,1 @@
+from .singleton import Singleton
